Tests/LibWeb/Text/expected/css/animate-with-many-keyframes.txt

rect.left > 0: true
rect.left <= 100: true
rect.top > 0: true
